requirements.txt
Django==3.1.3
psycopg2-binary>=2.8
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
django-debug-toolbar==3.1.1
django-recaptcha==2.0.6
html5lib==1.1
idna==2.10
lxml==4.6.1
Pillow==8.0.1
pytz==2020.4
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.26.2
webencodings==0.5.1
uWSGI>=2.0.18,<2.1
